Adding commands.py for clean code and importability
