+++ docker_test/VERSION
@@ -1,0 +1,4 @@
+Version: 1.0.1
+Released: 23 August 2024
+
+# License and Changelog at https://github.com/untergeek/es-docker-test-scripts
