+++ pytest.ini
@@ -1,3 +1,5 @@
 [pytest]
 #log_cli=true
+env_files =
+    .env
 log_format = %(asctime)s %(levelname)-9s %(name)22s %(funcName)22s:%(lineno)-4d %(message)s
